Move files into components folder
