+++ src/components/Fairing.tsx
